+++ src/skyscraper.clj
@@ -116,8 +116,8 @@
 (defn processor
   "Performs a single stage of scraping."
   [input-context
-   {:keys [html-cache processed-cache update http-options retries]
-    :or {html-cache true, processed-cache true, update false, http-options nil, retries 5}}
+   {:keys [html-cache processed-cache update http-options retries cache-key-callback processor-name]
+    :or {html-cache true, processed-cache true, update false, http-options nil, retries 5, cache-key-callback (constantly nil)}}
    &
    {:keys [url-fn cache-key-fn cache-template process-fn updatable]
     :or {url-fn :url}}]
@@ -126,6 +126,7 @@
         cache-key-fn (or cache-key-fn #(format-template cache-template %))
         cache-key (cache-key-fn input-context)
         force (and update updatable)]
+    (cache-key-callback processor-name cache-key)
     (or
       (when-not force
         (cache/load processed-cache cache-key))
@@ -141,9 +142,9 @@
         processed))))
 
 (defmacro defprocessor
-  [name & opts]
-  `(defn ~name [context# user-opts#]
-     (processor context# user-opts# ~@opts)))
+  [processor-name & opts]
+  `(defn ~processor-name [context# user-opts#]
+     (processor context# (assoc user-opts# :processor-name ~(keyword (name processor-name))) ~@opts)))
 
 ;;; Scraping engine
 
@@ -186,6 +187,19 @@
     (with-open [f (io/writer output)]
       (csv/write-csv f [(map name ks)])
       (csv/write-csv f (map (fn [row-data] (map (comp str row-data) ks)) (do-scrape data (assoc params :update false)))))))
+
+;; FIXME: Callbacks and mutable state are a quick-and-dirty way of
+;; writing this function without rewriting the rest of the code too
+;; much, but they have many drawbacks. I'd like this function to
+;; be lazy and pure.
+(defn get-cache-keys
+  "Runs scraping on data and params, but returns a vector of cache keys
+  along with processor names that generated them."
+  [data & {:as params}]
+  (let [result (atom [])
+        cache-key-callback #(swap! result conj {:processor %1, :key %2})]
+    (dorun (do-scrape data (assoc params :cache-key-callback cache-key-callback)))
+    @result))
 
 (defn separate
   [pred coll]